requirements.txt
# Core AI/ML dependencies
torch==2.4.0
numpy==1.26.4
onnx==1.17.0
onnxruntime-gpu==1.20.1
insightface==0.7.3
tensorflow==2.18.0
keras
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pillow==10.2.0
scipy==1.15.0

# Web & Database dependencies
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
pymongo==4.6.1
requests==2.31.0
Werkzeug==3.0.1
SQLAlchemy==2.0.27

# Python utilities
protobuf==4.25.2
psutil==6.1.1
tqdm==4.67.1
pyyaml==6.0.2
urllib3==2.2.0
pyparsing==3.1.4

# Optional GUI dependencies (can remove if running headless)
customtkinter==5.2.2
tkinterdnd2==0.4.2
